refactor(state): extract chainstate manager to chainstate.rs
Move ChainstateManager and ChainstateManagerOptions to dedicated
chainstate.rs module.
