+++ src/state/chainstate.rs
@@ -1,0 +1,214 @@
+use std::ffi::CString;
+
+use libbitcoinkernel_sys::{
+    btck_BlockHash, btck_ChainstateManager, btck_ChainstateManagerOptions, btck_block_read,
+    btck_block_spent_outputs_read, btck_chainstate_manager_create, btck_chainstate_manager_destroy,
+    btck_chainstate_manager_get_active_chain, btck_chainstate_manager_get_block_tree_entry_by_hash,
+    btck_chainstate_manager_import_blocks, btck_chainstate_manager_options_create,
+    btck_chainstate_manager_options_destroy,
+    btck_chainstate_manager_options_set_block_tree_db_in_memory,
+    btck_chainstate_manager_options_set_chainstate_db_in_memory,
+    btck_chainstate_manager_options_set_wipe_dbs,
+    btck_chainstate_manager_options_set_worker_threads_num, btck_chainstate_manager_process_block,
+};
+
+use crate::{
+    ffi::c_helpers, Block, BlockHash, BlockSpentOutputs, BlockTreeEntry, Chain, KernelError,
+};
+
+use super::Context;
+
+/// The chainstate manager is the central object for doing validation tasks as
+/// well as retrieving data from the chain. Internally it is a complex data
+/// structure with diverse functionality.
+///
+/// The chainstate manager is only valid for as long as the [`Context`] with which it
+/// was created remains in memory.
+///
+/// Its functionality will be more and more exposed in the future.
+pub struct ChainstateManager {
+    inner: *mut btck_ChainstateManager,
+}
+
+unsafe impl Send for ChainstateManager {}
+unsafe impl Sync for ChainstateManager {}
+
+impl ChainstateManager {
+    pub fn new(chainman_opts: ChainstateManagerOptions) -> Result<Self, KernelError> {
+        let inner = unsafe { btck_chainstate_manager_create(chainman_opts.inner) };
+        if inner.is_null() {
+            return Err(KernelError::Internal(
+                "Failed to create chainstate manager.".to_string(),
+            ));
+        }
+        Ok(Self { inner })
+    }
+
+    /// Process and validate the passed in block with the [`ChainstateManager`]
+    /// If processing failed, some information can be retrieved through the status
+    /// enumeration. More detailed validation information in case of a failure can
+    /// also be retrieved through a registered validation interface. If the block
+    /// fails to validate the `block_checked` callback's ['BlockValidationState'] will
+    /// contain details.
+    pub fn process_block(&self, block: &Block) -> (bool /* accepted */, bool /* duplicate */) {
+        let mut new_block: i32 = 0;
+        let accepted = unsafe {
+            btck_chainstate_manager_process_block(self.inner, block.as_ptr(), &mut new_block)
+        };
+        (c_helpers::success(accepted), c_helpers::enabled(new_block))
+    }
+
+    /// May be called after load_chainstate to initialize the
+    /// [`ChainstateManager`]. Triggers the start of a reindex if the option was
+    /// previously set for the chainstate and block manager. Can also import an
+    /// array of existing block files selected by the user.
+    pub fn import_blocks(&self) -> Result<(), KernelError> {
+        let result = unsafe {
+            btck_chainstate_manager_import_blocks(
+                self.inner,
+                std::ptr::null_mut(),
+                std::ptr::null_mut(),
+                0,
+            )
+        };
+        match c_helpers::success(result) {
+            true => Ok(()),
+            false => Err(KernelError::Internal(
+                "Failed to import blocks.".to_string(),
+            )),
+        }
+    }
+
+    /// Read a block from disk by its block tree entry.
+    pub fn read_block_data(&self, entry: &BlockTreeEntry) -> Result<Block, KernelError> {
+        let inner = unsafe { btck_block_read(self.inner, entry.as_ptr()) };
+        if inner.is_null() {
+            return Err(KernelError::Internal("Failed to read block.".to_string()));
+        }
+        Ok(unsafe { Block::from_ptr(inner) })
+    }
+
+    /// Read a block's spent outputs data from disk by its block tree entry.
+    pub fn read_spent_outputs(
+        &self,
+        entry: &BlockTreeEntry,
+    ) -> Result<BlockSpentOutputs, KernelError> {
+        let inner = unsafe { btck_block_spent_outputs_read(self.inner, entry.as_ptr()) };
+        if inner.is_null() {
+            return Err(KernelError::Internal(
+                "Failed to read undo data.".to_string(),
+            ));
+        }
+        Ok(unsafe { BlockSpentOutputs::from_ptr(inner) })
+    }
+
+    pub fn active_chain(&self) -> Chain<'_> {
+        let ptr = unsafe { btck_chainstate_manager_get_active_chain(self.inner) };
+        unsafe { Chain::from_ptr(ptr) }
+    }
+
+    pub fn get_block_tree_entry(&self, block_hash: &BlockHash) -> Option<BlockTreeEntry<'_>> {
+        let ptr = unsafe {
+            btck_chainstate_manager_get_block_tree_entry_by_hash(
+                self.inner,
+                block_hash as *const _ as *const btck_BlockHash,
+            )
+        };
+        if ptr.is_null() {
+            None
+        } else {
+            Some(unsafe { BlockTreeEntry::from_ptr(ptr) })
+        }
+    }
+}
+
+impl Drop for ChainstateManager {
+    fn drop(&mut self) {
+        unsafe {
+            btck_chainstate_manager_destroy(self.inner);
+        }
+    }
+}
+
+/// Holds the configuration options for creating a new [`ChainstateManager`]
+pub struct ChainstateManagerOptions {
+    inner: *mut btck_ChainstateManagerOptions,
+}
+
+impl ChainstateManagerOptions {
+    /// Create a new option
+    ///
+    /// # Arguments
+    /// * `context` -  The [`ChainstateManager`] for which these options are created has to use the same [`Context`].
+    /// * `data_dir` - The directory into which the [`ChainstateManager`] will write its data.
+    pub fn new(context: &Context, data_dir: &str, blocks_dir: &str) -> Result<Self, KernelError> {
+        let c_data_dir = CString::new(data_dir)?;
+        let c_blocks_dir = CString::new(blocks_dir)?;
+        let inner = unsafe {
+            btck_chainstate_manager_options_create(
+                context.as_ptr(),
+                c_data_dir.as_ptr(),
+                c_data_dir.as_bytes().len(),
+                c_blocks_dir.as_ptr(),
+                c_blocks_dir.as_bytes().len(),
+            )
+        };
+        if inner.is_null() {
+            return Err(KernelError::Internal(
+                "Failed to create chainstate manager options.".to_string(),
+            ));
+        }
+        Ok(Self { inner })
+    }
+
+    /// Set the number of worker threads used by script validation
+    pub fn set_worker_threads(&self, worker_threads: i32) {
+        unsafe {
+            btck_chainstate_manager_options_set_worker_threads_num(self.inner, worker_threads);
+        }
+    }
+
+    /// Wipe the block tree or chainstate dbs. When wiping the block tree db the
+    /// chainstate db has to be wiped too. Wiping the databases will triggere a
+    /// rebase once import blocks is called.
+    pub fn set_wipe_db(self, wipe_block_tree: bool, wipe_chainstate: bool) -> Self {
+        unsafe {
+            btck_chainstate_manager_options_set_wipe_dbs(
+                self.inner,
+                c_helpers::to_c_bool(wipe_block_tree),
+                c_helpers::to_c_bool(wipe_chainstate),
+            );
+        }
+        self
+    }
+
+    /// Run the block tree db in-memory only. No database files will be written to disk.
+    pub fn set_block_tree_db_in_memory(self, block_tree_db_in_memory: bool) -> Self {
+        unsafe {
+            btck_chainstate_manager_options_set_block_tree_db_in_memory(
+                self.inner,
+                c_helpers::to_c_bool(block_tree_db_in_memory),
+            );
+        }
+        self
+    }
+
+    /// Run the chainstate db in-memory only. No database files will be written to disk.
+    pub fn set_chainstate_db_in_memory(self, chainstate_db_in_memory: bool) -> Self {
+        unsafe {
+            btck_chainstate_manager_options_set_chainstate_db_in_memory(
+                self.inner,
+                c_helpers::to_c_bool(chainstate_db_in_memory),
+            );
+        }
+        self
+    }
+}
+
+impl Drop for ChainstateManagerOptions {
+    fn drop(&mut self) {
+        unsafe {
+            btck_chainstate_manager_options_destroy(self.inner);
+        }
+    }
+}
