+++ src/state/mod.rs
@@ -1,3 +1,5 @@
+pub mod chainstate;
 pub mod context;
 
+pub use chainstate::{ChainstateManager, ChainstateManagerOptions};
 pub use context::{ChainParams, ChainType, Context, ContextBuilder};
